Changed OscilNode partial to support discrete or continuous. Updated GPExperiment to beginning adding continuous partial support

+++ Source/Synth/Nodes/OscilNode.h
@@ -28,7 +28,7 @@ public:
 
 private:
     bool terminalOscil;
-    int partial;
+    float partial;
     int variableNum;
     double w;
     double index;
